+++ test/physics/albedo.jl
@@ -1,20 +1,23 @@
 import Statistics: mean
 
 @testset "Single Albedos" begin
-    for Model in (PrimitiveWetModel, PrimitiveDryModel)
+    @testset for Model in (PrimitiveWetModel, PrimitiveDryModel)
         spectral_grid = SpectralGrid(trunc=31, nlayers=8)
 
-        for AlbedoType in (GlobalConstantAlbedo, ManualAlbedo, AlbedoClimatology)
+        @testset for AlbedoType in (GlobalConstantAlbedo, ManualAlbedo, AlbedoClimatology)
 
             albedo = AlbedoType(spectral_grid)
             AlbedoType == ManualAlbedo && set!(albedo, 0.06)
 
             model = Model(spectral_grid; albedo)
-            simulation = initialize!(model)
-            run!(simulation, steps=1)
 
-            a, b = extrema(simulation.diagnostic_variables.physics.albedo)
-            m = mean(simulation.diagnostic_variables.physics.albedo)
+            initialize!(albedo, model)
+            progn = PrognosticVariables(model)
+            diagn = DiagnosticVariables(model)
+            SpeedyWeather.parameterization_tendencies!(diagn, progn, model)
+
+            a, b = extrema(diagn.physics.albedo)
+            m = mean(diagn.physics.albedo)
             @test a >= 0
             @test b <= 1
             @test 0 < m < 1
@@ -26,30 +29,31 @@ end
     for Model in (PrimitiveWetModel, PrimitiveDryModel)
         spectral_grid = SpectralGrid(trunc=31, nlayers=8)
 
-        for AlbedoType1 in (GlobalConstantAlbedo, ManualAlbedo, AlbedoClimatology)
-            for AlbedoType2 in (GlobalConstantAlbedo, ManualAlbedo, AlbedoClimatology)
+        @testset for OceanAlbedo in (GlobalConstantAlbedo, ManualAlbedo, AlbedoClimatology, OceanSeaIceAlbedo)
+            @testset for LandAlbedo in (GlobalConstantAlbedo, ManualAlbedo, AlbedoClimatology, LandSnowAlbedo)
 
-                albedo = Albedo(
-                    ocean=AlbedoType1(spectral_grid),
-                    land=AlbedoType2(spectral_grid))
+                albedo = OceanLandAlbedo(
+                    ocean=OceanAlbedo(spectral_grid),
+                    land=LandAlbedo(spectral_grid))
 
-                AlbedoType1 == ManualAlbedo && set!(albedo.ocean, 0.06)
-                AlbedoType2 == ManualAlbedo && set!(albedo.land, 0.06)
+                OceanAlbedo == ManualAlbedo && set!(albedo.ocean, 0.06)
+                LandAlbedo == ManualAlbedo && set!(albedo.land, 0.06)
 
                 model = Model(spectral_grid; albedo)
-                simulation = initialize!(model)
 
-                progn, diagn, model = SpeedyWeather.unpack(simulation)
-                SpeedyWeather.albedo!(diagn, progn, model)
+                initialize!(albedo, model)
+                progn = PrognosticVariables(model)
+                diagn = DiagnosticVariables(model)
+                SpeedyWeather.parameterization_tendencies!(diagn, progn, model)
 
-                a, b = extrema(simulation.diagnostic_variables.physics.land.albedo)
-                m = mean(simulation.diagnostic_variables.physics.land.albedo)
+                a, b = extrema(diagn.physics.land.albedo)
+                m = mean(diagn.physics.land.albedo)
                 @test a >= 0
                 @test b <= 1
                 @test 0 < m < 1
 
-                a, b = extrema(simulation.diagnostic_variables.physics.ocean.albedo)
-                m = mean(simulation.diagnostic_variables.physics.ocean.albedo)
+                a, b = extrema(diagn.physics.ocean.albedo)
+                m = mean(diagn.physics.ocean.albedo)
                 @test a >= 0
                 @test b <= 1
                 @test 0 < m < 1