+++ test/physics/convection.jl
@@ -2,15 +2,16 @@
     
     spectral_grid = SpectralGrid(trunc=31, nlayers=8)
 
-    for Convection in ( Nothing,
-                        SimplifiedBettsMiller,
-                        DryBettsMiller,
-                        ConvectiveHeating)
-        for Model in (  PrimitiveDryModel,
-                        PrimitiveWetModel)
+    @testset for Convection in ( Nothing,
+                                BettsMillerConvection,
+                                BettsMillerDryConvection,
+                                ConvectiveHeating)
+            
+        @testset for Model in (  PrimitiveDryModel,
+                                PrimitiveWetModel)
 
             # that combination is not defined
-            if ~(Convection == SimplifiedBettsMiller && Model == PrimitiveDryModel)
+            if ~(Convection == BettsMillerConvection && Model == PrimitiveDryModel)
                 convection = Convection(spectral_grid)
                 model = Model(spectral_grid; convection)
                 model.feedback.verbose = false